refactor path
