+++ app/dcr_api/biz/pkg/storage/storage.go
@@ -11,8 +11,6 @@ import (
 type Storage interface {
 	BucketPath() string
 	UploadFile(reader io.Reader, remotePath string, compress bool) error
-	GetFileSize(remotePath string) (int64, error)
-	GetFilebyChunk(remotePath string, offset int64, chunkSize int64) ([]byte, error)
 	PresignedUrl(remotePath string, method string, expiry time.Duration) (string, error)
 	Close()
 }
