+++ app/dcr_api/biz/service/job_service.go
@@ -27,15 +27,14 @@ import (
 	"github.com/google/uuid"
 	"github.com/manatee-project/manatee/app/dcr_api/biz/dal/db"
 	"github.com/manatee-project/manatee/app/dcr_api/biz/model/job"
+	"github.com/manatee-project/manatee/app/dcr_api/biz/pkg/storage"
 	"github.com/manatee-project/manatee/pkg/errno"
-	"github.com/manatee-project/manatee/pkg/storage"
-	bucketStorage "github.com/manatee-project/manatee/pkg/storage"
 	"github.com/pkg/errors"
 )
 
 type JobService struct {
 	ctx     context.Context
-	storage bucketStorage.Storage
+	storage storage.Storage
 }
 
 // NewJobService create job service
