+++ app/dcr_monitor/registry/registry.go
